+++ magda-minion-format/src/index.ts
@@ -2,7 +2,6 @@ import minion from "@magda/minion-framework/dist/index";
 import onRecordFound from "./onRecordFound";
 import formatAspectDef from "./formatAspectDef";
 import commonYargs from "@magda/minion-framework/dist/commonYargs";
-import { MAGDA_SYSTEM_ID } from "@magda/typescript-common/dist/registry/TenantConsts";
 
 const ID = "minion-format";
 
@@ -16,7 +15,6 @@ function sleuthLayerer() {
         optionalAspects: [],
         async: false,
         writeAspectDefs: [formatAspectDef],
-        tenantId: MAGDA_SYSTEM_ID,
         onRecordFound
     });
 }