+++ magda-web-client/src/Components/Common/TooltipWrapper.tsx
@@ -1,22 +1,41 @@
-import "./Tooltip.scss";
-
 import React from "react";
+
+import "./TooltipWrapper.scss";
+
+type Props = {
+    /** Does this require a click outside the tooltip to dismiss, or should it disappear after a delay?  */
+    requireClickToDismiss?: boolean;
+    /** Invoked when the tooltip is dismissed by the user - not called if the tooltip disappears automatically  */
+    onDismiss?: () => void;
+    /** Classname to apply to the wrapper element */
+    className?: string;
+    /** Whether the tooltip should start in an open state */
+    startOpen?: boolean;
+    /** Whether the tooltip should show up above or below the element it wraps */
+    orientation?: "below" | "above";
+    /** The wrapped component that should launch the tooltip. If startOpen is false then it will show the tooltip when hovered over */
+    launcher?: React.ComponentType;
+    /** Class to apply to the  actual tooltip */
+    innerElementClassName?: string;
+    /** The tooltip content itself, as higher-order function that provides a function to dismiss the tooltip */
+    children: (dismiss: () => void) => React.ReactNode;
+};
+
+type State = {
+    offset: number;
+    dismissed: boolean;
+};
 
 /**
  * @description Return a information tooltip, on hover show calculation method.
- * @returns { div }
  */
-class Tooltip extends React.Component {
-    constructor(props) {
-        super(props);
-
-        this.rootRef = React.createRef();
-        this.tooltipTextElementRef = React.createRef();
-        this.state = {
-            offset: 0,
-            dismissed: false
-        };
-    }
+class TooltipWrapper extends React.Component<Props, State> {
+    rootRef = React.createRef<HTMLDivElement>();
+    tooltipTextElementRef = React.createRef<HTMLSpanElement>();
+    state = {
+        offset: 0,
+        dismissed: false
+    };
 
     componentDidMount() {
         if (this.props.requireClickToDismiss) {
@@ -36,7 +55,10 @@ class Tooltip extends React.Component {
     }
 
     handleClickOutside = event => {
-        if (!this.rootRef.current.contains(event.target)) {
+        if (
+            !this.rootRef.current ||
+            !this.rootRef.current.contains(event.target)
+        ) {
             this.dismiss();
         }
     };
@@ -54,13 +76,13 @@ class Tooltip extends React.Component {
         const rootElement = this.rootRef.current;
 
         // Why .firstChild? Because we can't attach a ref to a render prop unless whatever's passed in passes the ref through to its first dom element
-        const launcherElement = rootElement.firstChild;
+        const launcherElement = rootElement!.firstChild!;
 
         const launcherElementStyle =
-            launcherElement.currentStyle ||
-            window.getComputedStyle(launcherElement);
+            (launcherElement as any).currentStyle ||
+            window.getComputedStyle(launcherElement as Element);
 
-        const tooltipWidth = tooltipTextElement.offsetWidth;
+        const tooltipWidth = tooltipTextElement!.offsetWidth;
         const offset =
             (tooltipWidth +
                 parseFloat(launcherElementStyle.marginLeft) +
@@ -107,4 +129,4 @@ class Tooltip extends React.Component {
     }
 }
 
-export default Tooltip;
+export default TooltipWrapper;