+++ CHANGES.md
@@ -20,6 +20,7 @@ Gateway:
 -   Add tenant ID header to client requests.
 -   Add ArcGIS/ESRI Authentication provider, including support for on-premise instances of ArcGIS Portal.
 -   Add Vanguard (WS-FED) Authentication provider
+-   Add `webProxyRoutesJson` command-line argument, allowing non-API proxy routes to be configured.
 
 Search:
 
@@ -89,6 +90,8 @@ UI:
 -   Show an error message screen if the user is not allowed to access the add dataset page
 -   Updated security classification & sensitivity questions according to the new design
 -   Make sure all publish new dataset errors are captured and shown to users on add dataset page
+-   Added Custodian field to the 'People and Production' page
+-   Added the ability to add references to datasets that a new dataset was derived from.
 
 Gateway:
 
@@ -122,6 +125,7 @@ Others:
 -   Format minion will trust dcat format if other measures indicate a ESRI REST format
 -   Added ASC to 4 stars rating list
 -   Removed Travis CI (Gitlab CI still remains)
+-   Disabled tenant-api & tenant-db when `enableMultiTenants` = false
 
 ## 0.0.55